// src/services/types.ts
export interface ParsedCNPJ {
  cnpj?: string;
  razaoSocial?: string;
  nomeFantasia?: string;
  dataAbertura?: string;
  logradouro?: string;
  numero?: string;
  complemento?: string;
  cep?: string;
  bairro?: string;
  municipio?: string;
  uf?: string;
  email?: string;
  telefone?: string;
  porte?: string;
  naturezaJuridica?: string;
  cnaePrincipal?: string;
  situacaoCadastral?: string;
}